requirements.txt
fastapi[all]
supabase
python-dotenv
outscraper